chore: renamed action maps to singlular
