+++ apps/web/src/app/(dashboard)/(account)/profiel/[handle]/_components/certificate/add-certificate.tsx
@@ -5,7 +5,7 @@ import { useAction } from "next-safe-action/hooks";
 import { useState } from "react";
 import { useFormStatus } from "react-dom";
 import { toast } from "sonner";
-import { createExternalCertificateAction } from "~/actions/certficates/create-external-certificate-action";
+import { createExternalCertificateAction } from "~/actions/certficate/create-external-certificate-action";
 import { Button } from "~/app/(dashboard)/_components/button";
 import {
   Dialog,
